PinchHandler: don't confine activeScale and activeRotation to min/max
It was inconsistent to do this for the active values only on one
handler's axis objects. The main purpose of the minimum and maximum
values is to avoid modifying the target item's respective properties
beyond reasonable limits, whereas the active values are supposed to be
derived directly from the gesture. If the users modify target properties
based on active values, or based on the activeValueChanged signal
argument, they need to enforce the property limits themselves.

[ChangeLog][QtQuick][Event Handlers] PinchHandler's activeScale (which
was previously called scale) is no longer restricted to the range
between minimumScale and maximumScale: these limits apply only to
persistentScale. Likewise, activeRotation (previously rotation)
is no longer restricted to the range between minimumRotation and
maximumRotation.

Change-Id: I5757e3d931e6165c81cecff19fc30b50033d89ce
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>