Stabilize rendering of Rectangle borders under fractional scaling
By default, border widths are rounded to integer values (in
QSGBasicInternalRectangleNode::updateGeometry()), so one avoids
artifacts where the rendered width can fluctuate with one pixel if the
Rectangle coordinates are fractional. However, that rounding was done
before the device pixel ratio scaling, so if the dpr was fractional,
one would get a fractional rendered border width, and hence
fluctuations.

Fix by taking the dpr into account. The rounding is done in the Item,
where the dpr is known, instead of the SG Node.

Fixes: QTBUG-108831
Pick-to: 6.4 6.2
Change-Id: I9a1d8180c72dd5e1b1eaa9f1c420eb9944f1e5a5
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>