Add containerStyle property to Material attached object
The Material 3 TextField has 2 different variants.
It defines one of those variants as the "Filled" variant, which looks
similar to the Material 2 version.
But it also defines an "Outlined" variant, which looks very different.

The Material 3 buttons also share a similar concept.

This patch introduces a new property to the attached object, which we
can use to implement the 2 different variants of TextField, and other
controls with similar filled/outlined distinctions.

[ChangeLog][Controls][Material] Added containerStyle attached property,
which can be used on certain controls to change the style of the
container that they are rendered in.

Task-number: QTBUG-109218
Change-Id: Ic468863be14c0c44981411c8f3dbea5d1d0107b0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>