QQuickTableView: never start to edit from a modifier key alone
When TableView.editTriggers is set to AnyKeyPressed, the
user can start to edit the current cell by pressing any
key on the keyboard. But there are some exceptions, like
the arrow keys. But it turns out that we also need to
ignore the modifier keys, otherwise the user cannot
e.g do a shift-drag to select cells without also
starting to edit.

Change-Id: I26ca954349823b50a605908633eb9c1d4996076a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>