Make the ShaderEffect type cache self maintaining
...meaning preventing infinite growth with ShaderEffects that
continuously change their shaders to something not encountered
before, while removing the explicit cleanup hooks that pollute
the render loops and similar. All it needs is a simple reference
counting and ensuring the material derefs whenever it goes away
or when it changes the associated shaders.

This way one does not even need a QQuickWindow::releaseResources()
to ensure a minimal type cache, it stays minimal (has only active
entries) by design. (unlike other caches the data here is not
valuable, so when all effects with the same shader sources go away,
there is no point in keeping that material key, we can easily
generate a new one if the same shader set pops up later)

Change-Id: I9ddf976960b0e4ecae22d9a5cca1a7a4faa4eb9e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Christian Strømme <christian.stromme@qt.io>