Doc: Fix incorrect \inqmlmodule statements
CloseEvent: Refers to an undocumented module QtQuick.Window which
is now an alias to QtQuick import.

ImputMethod: Trailing full stop in the module name.

These were not caught by documentation testing in the CI because of
a bug in QDoc where \inqmlmodule <module> did not output a warning
if <module> was undocumented.

Pick-to: 6.4
Change-Id: Ifd087f6d2ec5c32bd69700b7b131581895da26d0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>