Fix ListView.isCurrentItem for DelegateModel, take 2
The attached ListView.isCurrentItem property is updated when
QQuickItemViewPrivate::updateCurrent() method is called.
During the initialization it could be potentially called twice:
* from the QQuickItemView::setCurrentIndex() method, but that does
  not happen, because the isComponentCompleted() condition is not
  fulfilled at that time.
* from QQuickItemView::componentComplete() method, but that does
  not happen, because the d->isValid() condition is not fulfilled.
  The latter means that the model is not yet initialized.

This patch attempts to fix it by adding the updateCurrent() call
into layout() method. This method is called each time the
component needs to be redrawn, so it creates the proper
currentItem during the first call. Most of the subsequent calls
will do nothing, because the currentItem will be non-null, and
the currentIndex will not change.

The unit-test is taken from the reverted commit
d9f9d773e92940786f159897623618f3bf6bcf0f.

Another unit-test is meant to check that there is no regression
like in QTBUG-98315.

Done-with: Joni Poikelin <joni.poikelin@qt.io>
Fixes: QTBUG-86744
Pick-to: 6.5 6.4 6.2
Change-Id: Iab86a9c0e22660126f152727e8bd393ac17f5d15
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>