Remove QQuickDragAxis::persistentValue and changed signal for now
It's better to avoid releasing with this new property until we are sure
we know its proper meaning. So far it seems that it should end up being
identical to the target-item property to be set, rather than e.g.
DragHandler.xAxis.persistentValue being the same as
DragHandler.persistentTranslation.x. That is, it's an item property's
current value, not a delta to adjust the value.

Reverts part of 7867a683fcb938939fb2837a26ac8e1941e3fe08

Pick-to: 6.5
Task-number: QTBUG-109373
Change-Id: Id3a78332194c714e75b69cbaef111229d75318e0
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>