QQuickSelectionRectangle: disable tapping through the selection handles
We need to create a dummy TapHandler on top of the selection
handle to stop the user from able to trigger a TapHandler
on an item underneath it.

The default gesture policy will stop a TapHandler underneath
from triggering, but it will not stop it from emitting
"pressed". We therefore also need to set a policy on the
handler that let's it take an exclusive grab.

Pick-to: 6.5 6.4
Change-Id: Ia49a4632b25b6c31a9c258f92880b95ad343e020
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>