Native style: remove setting font from QStyle
This is a follow-up on 43eca45b06.

Which font a control should use should follow
normal font propagation. So setting a different
font explicitly in every control is not ideal.
The font is in practice the same for all controls
anyway.

This will also remove a warning that will be
printed if the application assigns a custom
background delegate to a Button.

Fixes: QTBUG-98098
Change-Id: I4f9304b391b4b9e398882fc532cbff5b09f53aff
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>