QML: Re-allow assigning QVariantList to arbitrary other lists
Since we can assign a QVariant to any other singular value, we should be
able to do the same with lists of QVariant. This ability was lost when
we stopped converting QVariantList to JS array but rather started to
express it as QV4::Sequence.

As a drive-by, generalize the messy code that allows assigning singular
values to lists. This should also just work for any singular value and
any kind of list.

Fixes: QTBUG-109584
Change-Id: Icab1dceff332c5fc4fae9e7e8a8b5fde9f0b0ea9
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 32a9015c8f511608d98a9fdf16795b32c52052c3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>