QQuickTableView: don't load any delegates if viewport has zero size
When TableView has a zero size, the viewport will have zero
size as well, and there is no reason to load any delegates.

This patch will ensure that we return early from
loadInitialTable() when we detect that this is the case.

This will also stop a warning from being printed when
the delegate has it's implicitWidth bound to
tableView.width, and tableView.width == 0.

Pick-to: 6.5
Change-Id: I659eb098f7c5c25c8359876f7664499381f94b98
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>