QQuickTreeView: change the order of row and column to modelIndex(), v2
This patch amends 46fb848411. That change swapped the
order of row and column to QQuickTableView::modelIndex().
But that function is virtual, so we need to apply the
same change to the overridden function in QQuickTreeView
as well.

Task-number: QTBUG-109542
Change-Id: I8a2abbc16baba550683d6184573aecaad72c818d
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>
(cherry picked from commit 6bfc50dcb43c222af21d2c23eff909b57f4948e0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>