De-inline QQmlPrivate::revisionClassInfos()
It was compiled 1152 times, with the QList::push_back alone
contributing a cumulative ~24sec to a 'ninja qtdeclarative'. There's
no reason it should be inline, so make it out-of-line.

This function was the highest non-QMetaType/non-std contributor in a
Clang -ftime-trace build of qtdeclarative.

Task-number: QTBUG-97601
Change-Id: I143639168f859dc604b902647fe921332bbe49c0
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>
(cherry picked from commit 1e852bb496b93908f24869631443f1c971f2ce09)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>