De-inline DomItem::subReferencesItem()
It calls fromQList<Path>(), which is compiled 55x at almost 0.5s each
time round, contributing ~24s to a 'ninja qtdeclarative' build.

It is the top-most non-QMetaType/non-std template instantation
according to Clang -ftime-trace.

Also de-inlined DomItem::subReferenceItem() before realizing that
there are tons of other member functions that need to be de-inlined in
that class and giving up for this round.

Pick-to: 6.5
Task-number: QTBUG-97601
Change-Id: I23489c68b0a5b5df3c8bd0e4f6c3cb3c1ba5bd4f
Reviewed-by: Giuseppe D'Angelo <giuseppe.dangelo@kdab.com>