tst_QQuickListView::flickBothDirections: flick more
If we make flickDeceleration's default value platform-dependent and
often larger, this test begins failing because it expects to flick
all the way to the end of the ListView. If we flick a longer distance
(starting at the lower-right instead of in the middle of the ListView)
in the same time, it's more likely to get to the end.

An alternative would be to set flickDeceleration back to the old value
in this test.

Change-Id: If85d9502ceef9fbf6fc087240572cc98326453a5
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 0461a51c8e288ed64076fcfdbef604f1355b84a7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>